# feat: return body as consumable body2