--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -39,12 +39,7 @@ module.exports = class BodyReadable extends Readable {
 
   push (val) {
     if (this[kBody]) {
-      try {
-        return this[kBody].push(val)
-      } catch (err) {
-        this.destroy(err)
-        return false
-      }
+      return this[kBody].push(val)
     }
 
     return Readable.prototype.push.call(this, val)
@@ -128,20 +123,12 @@ function start (self) {
 }
 
 function consume (self, type) {
-  try {
-    if (self.bodyUsed) {
-      throw new TypeError('disturbed')
-    }
+  if (self.bodyUsed) {
+    throw new TypeError('disturbed')
+  }
 
-    if (self[kBody]) {
-      throw new TypeError('locked')
-    }
-  } catch (err) {
-    if (!type) {
-      self.destroy(err)
-    } else {
-      throw err
-    }
+  if (self[kBody]) {
+    throw new TypeError('locked')
   }
 
   if (!type) {
@@ -170,6 +157,10 @@ function consume (self, type) {
             controller,
             ended: false,
             push (val) {
+              if (self.destroyed) {
+                return false
+              }
+
               // TODO (fix): This is not strictly correct.
               this.used = true
 
@@ -183,9 +174,7 @@ function consume (self, type) {
                 this.ended = true
 
                 // TODO (fix): This is not strictly correct.
-                queueMicrotask(() => {
-                  Readable.prototype.push.call(self, null)
-                })
+                Readable.prototype.push.call(self, null)
               } else {
                 this.controller.enqueue(new Uint8Array(val))
               }
@@ -225,35 +214,42 @@ function consume (self, type) {
       ended: false,
       body: this.type === kTextType || this.type === kJSONType ? '' : [],
       push (val) {
+        if (self.destroyed) {
+          return false
+        }
+
         this.used = true
 
-        if (this.type === kTextType || this.type === kJSONType) {
-          if (val !== null) {
-            this.body += val
-          } else if (this.type === kTextType) {
-            resolve(this.body)
-          } else if (this.type === kJSONType) {
-            resolve(JSON.parse(this.body))
-          }
-        } else {
-          if (val !== null) {
-            this.body.push(val)
-          } else if (this.type === kArrayBufferType) {
-            resolve(Buffer.concat(this.body).buffer)
-          } else if (this.type === kBlobType) {
-            if (!Blob) {
-              Blob = require('buffer').Blob
+        try {
+          if (this.type === kTextType || this.type === kJSONType) {
+            if (val !== null) {
+              this.body += val
+            } else if (this.type === kTextType) {
+              resolve(this.body)
+            } else if (this.type === kJSONType) {
+              resolve(JSON.parse(this.body))
             }
-            resolve(new Blob(this.body))
+          } else {
+            if (val !== null) {
+              this.body.push(val)
+            } else if (this.type === kArrayBufferType) {
+              resolve(Buffer.concat(this.body).buffer)
+            } else if (this.type === kBlobType) {
+              if (!Blob) {
+                Blob = require('buffer').Blob
+              }
+              resolve(new Blob(this.body))
+            }
           }
+        } catch (err) {
+          self.destroy(err)
+          return false
         }
 
         if (val === null) {
           this.ended = true
           this.body = null
-          queueMicrotask(() => {
-            Readable.prototype.push.call(self, null)
-          })
+          Readable.prototype.push.call(self, null)
         }
 
         return true