--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -29,8 +29,7 @@ module.exports = class BodyReadable extends Readable {
   }
 
   destroy (err) {
-    // TODO (fix): This is not strictly correct.
-    if (!err && this[kBody] && !this[kBody].ended) {
+    if (this[kBody] && !err && !this._readableState.ended) {
       err = new AbortError()
     }
 
@@ -155,13 +154,14 @@ function consume (self, type) {
             used: false,
             buffer: self,
             controller,
-            ended: false,
             push (val) {
               if (self.destroyed) {
                 return false
               }
 
               // TODO (fix): This is not strictly correct.
+              // Just because chunk was enqueue doesn't mean
+              // that it was read?
               this.used = true
 
               if (!this.controller) {
@@ -171,10 +171,15 @@ function consume (self, type) {
 
               if (!val) {
                 this.controller.close()
-                this.ended = true
 
                 // TODO (fix): This is not strictly correct.
-                Readable.prototype.push.call(self, null)
+                // Just because chunk was enqueue doesn't mean
+                // that it was read? How do we wait for stream
+                // to be emptired? queueMicrotask is just a hack
+                // to hope that it's been drained.
+                queueMicrotask(() => {
+                  Readable.prototype.push.call(self, null)
+                })
               } else {
                 this.controller.enqueue(new Uint8Array(val))
               }
@@ -211,7 +216,6 @@ function consume (self, type) {
     self[kBody] = {
       type,
       used: false,
-      ended: false,
       body: this.type === kTextType || this.type === kJSONType ? '' : [],
       push (val) {
         if (self.destroyed) {
@@ -247,7 +251,6 @@ function consume (self, type) {
         }
 
         if (val === null) {
-          this.ended = true
           this.body = null
           Readable.prototype.push.call(self, null)
         }