--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -7,6 +7,7 @@ let Blob
 let ReadableStream
 
 const kBody = Symbol('body')
+const kDestroyed = Symbol('destroyed')
 
 const kWebStreamType = 1
 const kTextType = 2
@@ -26,12 +27,19 @@ module.exports = class BodyReadable extends Readable {
   constructor (opts) {
     super(opts)
     this[kBody] = undefined
+    this[kDestroyed] = false
   }
 
   destroy (err) {
+    if (this[kDestroyed]) {
+      return
+    }
+
     if (this[kBody] && !err && !this._readableState.ended) {
       err = new AbortError()
     }
+
+    this[kDestroyed] = true
 
     return Readable.prototype.destroy.call(this, err)
   }
@@ -146,9 +154,14 @@ function consume (self, type) {
           // TODO (fix): it's a little unclear what we need to do here.
           this.controller.error(new Error('locked'))
         } else {
-          self.on('error', err => {
-            this.controller.error(err)
-          })
+          self
+            .on('error', err => {
+              this.controller.error(err)
+            })
+            .on('end', () => {
+              // autoDestroy might have been disabled.
+              Readable.prototype.destroy.call(self, null)
+            })
           self[kBody] = {
             type,
             used: false,
@@ -160,7 +173,7 @@ function consume (self, type) {
               }
 
               // TODO (fix): This is not strictly correct.
-              // Just because chunk was enqueue doesn't mean
+              // Just because chunk was enqueued doesn't mean
               // that it was read?
               this.used = true
 
@@ -173,9 +186,9 @@ function consume (self, type) {
                 this.controller.close()
 
                 // TODO (fix): This is not strictly correct.
-                // Just because chunk was enqueue doesn't mean
+                // Just because chunk was enqueued doesn't mean
                 // that it was read? How do we wait for stream
-                // to be emptired? queueMicrotask is just a hack
+                // to be drained? queueMicrotask is just a hack
                 // to hope that it's been drained.
                 queueMicrotask(() => {
                   Readable.prototype.push.call(self, null)
@@ -212,7 +225,12 @@ function consume (self, type) {
   }
 
   return new Promise((resolve, reject) => {
-    self.on('error', reject)
+    self
+      .on('error', reject)
+      .on('end', () => {
+        // autoDestroy might have been disabled.
+        Readable.prototype.destroy.call(self, null)
+      })
     self[kBody] = {
       type,
       used: false,