--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -35,7 +35,7 @@ module.exports = class BodyReadable extends Readable {
       return
     }
 
-    if (this[kBody] && !err && !this._readableState.ended) {
+    if (this[kBody] && !err && !this._readableState.endEmitted) {
       err = new AbortError()
     }