--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -160,7 +160,7 @@ function consume (self, type) {
             })
             .on('end', () => {
               // autoDestroy might have been disabled.
-              Readable.prototype.destroy.call(self, null)
+              self.destroy()
             })
           self[kBody] = {
             type,
@@ -229,7 +229,7 @@ function consume (self, type) {
       .on('error', reject)
       .on('end', () => {
         // autoDestroy might have been disabled.
-        Readable.prototype.destroy.call(self, null)
+        self.destroy()
       })
     self[kBody] = {
       type,
@@ -242,18 +242,18 @@ function consume (self, type) {
 
         this.used = true
 
-        try {
+        if (val !== null) {
           if (this.type === kTextType || this.type === kJSONType) {
-            if (val !== null) {
-              this.body += val
-            } else if (this.type === kTextType) {
+            this.body += val
+          } else {
+            this.body.push(val)
+          }
+        } else {
+          try {
+            if (this.type === kTextType) {
               resolve(this.body)
             } else if (this.type === kJSONType) {
               resolve(JSON.parse(this.body))
-            }
-          } else {
-            if (val !== null) {
-              this.body.push(val)
             } else if (this.type === kArrayBufferType) {
               resolve(Buffer.concat(this.body).buffer)
             } else if (this.type === kBlobType) {
@@ -262,15 +262,12 @@ function consume (self, type) {
               }
               resolve(new Blob(this.body))
             }
-          }
-        } catch (err) {
-          self.destroy(err)
-          return false
-        }
 
-        if (val === null) {
-          this.body = null
-          Readable.prototype.push.call(self, null)
+            this.body = null
+            Readable.prototype.push.call(self, null)
+          } catch (err) {
+            self.destroy(err)
+          }
         }
 
         return true