--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -84,6 +84,10 @@ module.exports = class BodyReadable extends Readable {
     return this.on(ev, fn)
   }
 
+  get locked () {
+    return !!this[kBody]
+  }
+
   get bodyUsed () {
     if (this[kBody]) {
       return this[kBody].used
@@ -134,7 +138,7 @@ function consume (self, type) {
     throw new TypeError('disturbed')
   }
 
-  if (self[kBody]) {
+  if (self.locked) {
     throw new TypeError('locked')
   }