--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -158,14 +158,6 @@ function consume (self, type) {
           // TODO (fix): it's a little unclear what we need to do here.
           this.controller.error(new Error('locked'))
         } else {
-          self
-            .on('error', err => {
-              this.controller.error(err)
-            })
-            .on('end', () => {
-              // autoDestroy might have been disabled.
-              self.destroy()
-            })
           self[kBody] = {
             type,
             used: false,
@@ -205,7 +197,49 @@ function consume (self, type) {
             }
           }
         }
-        start(self)
+
+        return new Promise((resolve, reject) => {
+          // TODO (fix): This should wait until stream
+          // has been "constructed".
+
+          const orgEmit = self.emit
+
+          function cleanup () {
+            self.emit = orgEmit
+            self.off('error', check)
+          }
+
+          function check (err) {
+            if (err) {
+              cleanup()
+              reject(err)
+            } else if (this._readableState.constructed) {
+              self
+                .on('error', err => {
+                  this.controller.error(err)
+                })
+                .on('end', () => {
+                  // autoDestroy might have been disabled.
+                  self.destroy()
+                })
+
+              start(self)
+
+              cleanup()
+              resolve()
+            } else {
+              setTimeout(check, 100)
+            }
+          }
+
+          self.emit = function (ev, fn) {
+            check(null)
+            return orgEmit.call(this, ev, fn)
+          }
+          self.on('error', check)
+
+          check()
+        })
       },
 
       pull () {