--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -199,9 +199,6 @@ function consume (self, type) {
         }
 
         return new Promise((resolve, reject) => {
-          // TODO (fix): This should wait until stream
-          // has been "constructed".
-
           const orgEmit = self.emit
 
           function cleanup () {
@@ -210,10 +207,13 @@ function consume (self, type) {
           }
 
           function check (err) {
+            if (!err && !this._readableState.constructed) {
+              return
+            }
+
             if (err) {
-              cleanup()
               reject(err)
-            } else if (this._readableState.constructed) {
+            } else {
               self
                 .on('error', err => {
                   this.controller.error(err)
@@ -225,11 +225,10 @@ function consume (self, type) {
 
               start(self)
 
-              cleanup()
               resolve()
-            } else {
-              setTimeout(check, 100)
             }
+
+            cleanup()
           }
 
           self.emit = function (ev, fn) {