--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -84,10 +84,6 @@ module.exports = class BodyReadable extends Readable {
     return this.on(ev, fn)
   }
 
-  get locked () {
-    return !!this[kBody]
-  }
-
   get bodyUsed () {
     if (this[kBody]) {
       return this[kBody].used