--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -218,6 +218,10 @@ function consume (self, type) {
                   // autoDestroy might have been disabled.
                   self.destroy()
                 })
+                .on('close', () => {
+                  // TODO (fix): Do webstreams have something
+                  // corresponding?
+                })
 
               start(self)