--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -66,6 +66,13 @@ module.exports = class BodyReadable extends Readable {
     return Readable.prototype.resume.call(this)
   }
 
+  pause (dest, pipeOpts) {
+    if (this[kBody] === undefined) {
+      consume(this)
+    }
+    return Readable.prototype.pause.call(this)
+  }
+
   pipe (dest, pipeOpts) {
     if (this[kBody] === undefined) {
       consume(this)