--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -30,6 +30,16 @@ module.exports = class BodyReadable extends Readable {
     this[kDestroyed] = false
   }
 
+  get readableLength () {
+    // TODO (fix): Read from web stream?
+    return super.readableLength
+  }
+
+  get readableHighWaterMark () {
+    // TODO (fix): Read from web stream?
+    return super.readableHighWaterMark
+  }
+
   destroy (err) {
     if (this[kDestroyed]) {
       return