--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -165,6 +165,8 @@ function consume (self, type) {
       ReadableStream = require('stream/web').ReadableStream
     }
 
+    // TODO (fix): emit pause/resume on self?
+
     return new ReadableStream({
       start (controller) {
         if (self[kBody]) {