--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -25,6 +25,8 @@ class AbortError extends Error {
 
 module.exports = class BodyReadable extends Readable {
   constructor (opts) {
+    // TODO (fix): Improve allocation by making readableState
+    // allocated as lazily as possible and/or introduce pooling.
     super(opts)
     this[kBody] = undefined
     this[kDestroyed] = false