--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -94,6 +94,8 @@ module.exports = class BodyReadable extends Readable {
 
   on (ev, fn) {
     if (this[kBody] === undefined && (ev === 'data' || ev === 'readable')) {
+      // TODO (fix): consume in next tick in case event handler is removed
+      // in same tick?
       consume(this)
     }
     return Readable.prototype.on.call(this, ev, fn)