--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -60,7 +60,6 @@ module.exports = class BodyReadable extends Readable {
     if (this[kBody]) {
       return this[kBody].push(val)
     }
-
     return Readable.prototype.push.call(this, val)
   }
 
@@ -151,6 +150,11 @@ function start (self) {
 }
 
 function consume (self, type) {
+  if (self.destroyed) {
+    // TODO (fix): What error?
+    throw new TypeError('disturbed')
+  }
+
   if (self.bodyUsed) {
     throw new TypeError('disturbed')
   }