--- a/lib/api/readable.js
+++ b/lib/api/readable.js
@@ -159,7 +159,7 @@ function consume (self, type) {
     throw new TypeError('disturbed')
   }
 
-  if (self.locked) {
+  if (self[kBody]) {
     throw new TypeError('locked')
   }