This is an alternative variant solving some problems with the previous [PR](https://github.com/nodejs/undici/pull/895).

- How to return Node stream?
- Breaking change.

It's a little hacky but I think it can work out. Thoughts?