@ronag 2021-07-25
I've reduced the ambition here. Remove the hacky performance part and focus on API and functionality to begin with.