@ronag 2021-07-25
Also [opened against node core](https://github.com/nodejs/node/pull/39520). Waiting for feedback there as well.