@ronag 2021-07-28
I think it's likely we will land a better version in node core. https://github.com/nodejs/node/pull/39520